# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contoh API documentation",
        "description": "Contoh API doc menggunakan OpenAPI/Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "dwianggarans@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/gallery": {
            "get": {
                "tags": [
                    "gallery"
                ],
                "summary": "Returns a Sample API gallery response",
                "description": "A sample gallery to test out the API",
                "operationId": "gallery",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Successfully retrieved the gallery",
                                    "galleries": [
                                        {
                                            "id": 1,
                                            "title": "galllery 1",
                                            "description": "desc galllery 1",
                                            "picture": "672ad27a5fe841730859642.jpeg",
                                            "created_at": "2024-11-06T02:20:42.000000Z",
                                            "updated_at": "2024-11-06T02:20:42.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "detail": "strings"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/greet": {
            "get": {
                "tags": [
                    "greeting"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample greeting to test out the API",
                "operationId": "greet",
                "parameters": [
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "Nama depan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "Nama belakang",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil memproses masukan user",
                                    "data": {
                                        "output": "Halo John Doe",
                                        "firstname": "John",
                                        "lastname": "Doe"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Missing data: firstname and/or lastname are required."
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "gallery",
            "description": "gallery"
        },
        {
            "name": "greeting",
            "description": "greeting"
        }
    ]
}